# public/api_doc_fixed.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "官方网站后台 API",
        "version": "2.0.0",
        "description": "官方网站后台系统API文档"
    },
    "paths": {
        "/public-api/articles": {
            "get": {
                "summary": "获取公共文章列表",
                "tags": [
                    "公共接口"
                ],
                "parameters": [
                    {
                        "name": "type",
                        "description": "文章类型：news（新闻）或 wechat（公众号）",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "news",
                                "wechat"
                            ]
                        }
                    },
                    {
                        "name": "limit",
                        "description": "每页数量，默认20，最大100",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 100
                        }
                    },
                    {
                        "name": "page",
                        "description": "页码，默认1",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回文章列表",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "page": {
                                                    "type": "integer"
                                                },
                                                "limit": {
                                                    "type": "integer"
                                                },
                                                "pages": {
                                                    "type": "integer"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/public-api/news/{id}": {
            "get": {
                "summary": "获取新闻文章详情",
                "tags": [
                    "公共接口"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "description": "文章ID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回文章详情"
                    }
                }
            }
        },
        "/public-api/wechat/{id}": {
            "get": {
                "summary": "获取公众号文章详情",
                "tags": [
                    "公共接口"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "description": "文章ID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回文章详情"
                    }
                }
            }
        }
    }
}